pyproject.toml
[tool.poetry]
name = "news-analyzer"
version = "0.1.0"
description = ""
authors = ["Tatiana_Garbuzova <tatiannna0697@gmail.com>"]
readme = "README.md"
packages = [{include = "news_analyzer"}]

[tool.poetry.dependencies]
python = "^3.10"
matplotlib = "^3.9.2"
seaborn = "^0.13.2"
scikit-learn = "^1.5.1"
python-dotenv = "^1.0.1"

[tool.poetry.scripts]
run = "news_analyzer.scripts.run:main"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
